requirements.txt
pytz==2015.7
django-grappelli==2.7.2
